+++ InfinityHammer/behavior/CustomHealth.cs
@@ -1,7 +1,3 @@
-using System.Runtime.Remoting.Messaging;
-using TMPro;
-using Service;
-
 namespace InfinityHammer;
 
 public class CustomHealth